requirements.txt
kivymd>=1.2.0
kivy>=2.1.0
requests>=2.28.0
python-dotenv>=1.0.0
certifi>=2023.7.22
urllib3>=1.26.0
charset-normalizer>=2.0.0
idna>=2.10